requirements.txt
# installed system wide https://sipsimpleclient.org/installation/
python3-application==3.0.6
python3-eventlib==0.3.0
python3-gnutls==3.1.11
python3-msrplib==0.21.1
python3-otr==2.0.2
python3-sipsimple==5.3.0
python3-xcaplib==2.0.1

# pip installed
fastapi==0.111.0